LPC-4: renamed namespaces and folders
