+++ LoopstationCompanionApi.Tests/Helpers/JsonAsserts.cs
@@ -1,6 +1,6 @@
 ﻿using System.Text.Json;
 
-namespace LoopstationCompanionApi.UnitTests.Helpers;
+namespace LoopstationCompanionApi.Tests.Helpers;
 
 public static class JsonAsserts
 {
